PN-17353

+++ PAS_Database/dbo/Tables/Tables2/WorkOrderShipping.sql
@@ -63,7 +63,7 @@
     [soldAttention]                      VARCHAR (100)   NULL,
     [CustomerDomensticShippingShipViaId] BIGINT          NULL,
     [ShippingAccountInfo]                VARCHAR (200)   NULL,
-    [NoOfItems]                          INT             NULL,
+    [NoOfItems]                          DECIMAL (18, 6) NULL,
     [IsCustomerShipping]                 BIT             NULL,
     [IsManualShipping]                   BIT             NULL,
     [ManufactureCountryId]               INT             NULL,
@@ -88,6 +88,8 @@
     CONSTRAINT [FK_WorkOrderShipping_WorkOrder] FOREIGN KEY ([WorkOrderId]) REFERENCES [dbo].[WorkOrder] ([WorkOrderId]),
     CONSTRAINT [FK_WorkOrderShipping_WorkOrderPartNo] FOREIGN KEY ([WorkOrderPartNoId]) REFERENCES [dbo].[WorkOrderPartNumber] ([ID])
 );
+
+
